move database jobs methods to their own file
