+++ api.py
@@ -30,6 +30,7 @@ from fastchat.protocol.openai_api_protocol import (
     ErrorResponse,
 )
 
+from transformerlab.db.jobs import job_create, job_get_error_msg, job_update_status
 import transformerlab.db.session as db
 from transformerlab.routers import (
     data,
@@ -321,7 +322,7 @@ async def server_worker_start(
         json.dumps(inference_params),
     ]
 
-    job_id = await db.job_create(type="LOAD_MODEL", status="STARTED", job_data="{}", experiment_id=experiment_id)
+    job_id = await job_create(type="LOAD_MODEL", status="STARTED", job_data="{}", experiment_id=experiment_id)
 
     print("Loading plugin loader instead of default worker")
 
@@ -347,10 +348,10 @@ async def server_worker_start(
     if exitcode is not None and exitcode != 0:
         with open(dirs.GLOBAL_LOG_PATH, "a") as global_log:
             global_log.write(f"Error loading model: {model_name} with exit code {exitcode}\n")
-        error_msg = await db.job_get_error_msg(job_id)
+        error_msg = await job_get_error_msg(job_id)
         if not error_msg:
             error_msg = f"Exit code {exitcode}"
-            await db.job_update_status(job_id, "FAILED", error_msg)
+            await job_update_status(job_id, "FAILED", error_msg)
         return {"status": "error", "message": error_msg}
     with open(dirs.GLOBAL_LOG_PATH, "a") as global_log:
         global_log.write(f"Model loaded successfully: {model_name}\n")
