+++ test/db/test_db.py
@@ -11,29 +11,15 @@ from transformerlab.db.db import (
     experiment_get_by_name,
     get_plugins_of_type,
     get_training_template_by_name,
-    job_count_running,
     job_create_sync,
-    job_delete,
-    job_get_error_msg,
     job_mark_as_complete_if_running,
-    job_update,
     job_update_status_sync,
     job_update_sync,
-    jobs_get_next_queued_job,
     model_local_create,
     model_local_get,
     model_local_list,
     model_local_count,
     model_local_delete,
-    job_create,
-    job_get_status,
-    job_get,
-    job_update_status,
-    job_delete_all,
-    job_update_job_data_insert_key_value,
-    job_stop,
-    jobs_get_all,
-    jobs_get_all_by_experiment_and_type,
     experiment_get_all,
     experiment_create,
     experiment_get,
@@ -88,8 +74,25 @@ from transformerlab.db.workflows import (
     workflows_get_all,
     workflows_get_from_experiment,
     workflows_get_by_id,
-    workflows_get_from_experiment,
 )
+
+from transformerlab.db.jobs import (
+    job_create,
+    job_get_status,
+    job_get,
+    job_update_status,
+    job_delete_all,
+    job_update_job_data_insert_key_value,
+    job_stop,
+    jobs_get_all,
+    jobs_get_all_by_experiment_and_type,
+    jobs_get_next_queued_job,
+    job_count_running,
+    job_delete,
+    job_get_error_msg,
+    job_update,
+)
+
 
 from transformerlab.db.session import job_cancel_in_progress_jobs
 import transformerlab.db.session as db
