+++ transformerlab/routers/experiment/workflows.py
@@ -1,6 +1,7 @@
 from fastapi import APIRouter, UploadFile, Body
 from fastapi.responses import FileResponse
 import transformerlab.db.db as db
+import transformerlab.db.jobs as db_jobs
 from transformerlab.db.workflows import (
     workflow_count_queued,
     workflow_count_running,
@@ -355,7 +356,7 @@ async def workflow_runs_get_by_id(workflow_run_id: str, experimentId: int):
     try:
         job_ids = json.loads(workflow_run.get("job_ids", "[]"))
         for job_id in job_ids:
-            job = await db.job_get(job_id)
+            job = await db_jobs.job_get(job_id)
             if not job:
                 continue
 
@@ -410,7 +411,7 @@ async def cancel_workflow_run(workflow_run_id: str, experimentId: int):
     cancelled_jobs = []
 
     for job_id in current_job_ids:
-        await db.job_stop(job_id)  # This sets stop=True in job_data
+        await db_jobs.job_stop(job_id)  # This sets stop=True in job_data
         cancelled_jobs.append(job_id)
 
     # The workflow execution engine will automatically detect the stopped jobs
@@ -463,7 +464,7 @@ async def check_current_jobs_status(workflow_run_id, current_job_ids):
         return None
 
     for job_id in current_job_ids:
-        current_job = await db.job_get(job_id)
+        current_job = await db_jobs.job_get(job_id)
         if not current_job:
             await workflow_run_update_status(workflow_run_id, "FAILED")
             return f"Could not find job with ID {job_id}"
@@ -587,7 +588,7 @@ async def find_previous_node_and_job(current_node, workflow_run, workflow_config
             ran_jobs = json.loads(ran_jobs_str)
             if previous_node["id"] in ran_nodes:
                 previous_job_ID = ran_jobs[ran_nodes.index(previous_node["id"])]
-                previous_job = await db.job_get(previous_job_ID)
+                previous_job = await db_jobs.job_get(previous_job_ID)
 
     return previous_job
