+++ transformerlab/routers/jobs.py
@@ -6,7 +6,6 @@ import logging
 from fastapi import APIRouter, Body, Response
 from fastapi.responses import StreamingResponse, FileResponse
 
-import transformerlab.db.db as db
 from transformerlab.shared import shared
 from transformerlab.shared import dirs
 from typing import Annotated
@@ -16,53 +15,55 @@ from werkzeug.utils import secure_filename
 
 from transformerlab.routers.serverinfo import watch_file
 
+import transformerlab.db.jobs as db_jobs
+
 router = APIRouter(prefix="/jobs", tags=["train"])
 
 
 @router.get("/list")
 async def jobs_get_all(type: str = "", status: str = ""):
-    jobs = await db.jobs_get_all(type=type, status=status)
+    jobs = await db_jobs.jobs_get_all(type=type, status=status)
     return jobs
 
 
 @router.get("/delete/{job_id}")
 async def job_delete(job_id: str):
-    await db.job_delete(job_id)
+    await db_jobs.job_delete(job_id)
     return {"message": "OK"}
 
 
 @router.get("/create")
 async def job_create(type: str = "UNDEFINED", status: str = "CREATED", data: str = "{}", experiment_id: str = "-1"):
-    jobid = await db.job_create(type=type, status=status, job_data=data, experiment_id=experiment_id)
+    jobid = await db_jobs.job_create(type=type, status=status, job_data=data, experiment_id=experiment_id)
     return jobid
 
 
 async def job_create_task(script: str, job_data: str = "{}", experiment_id: str = "-1"):
-    jobid = await db.job_create(type="UNDEFINED", status="CREATED", job_data=job_data, experiment_id=experiment_id)
+    jobid = await db_jobs.job_create(type="UNDEFINED", status="CREATED", job_data=job_data, experiment_id=experiment_id)
     return jobid
 
 
 @router.get("/update/{job_id}")
 async def job_update(job_id: str, status: str):
-    await db.job_update_status(job_id, status)
+    await db_jobs.job_update_status(job_id, status)
     return {"message": "OK"}
 
 
 @router.get("/start_next")
 async def start_next_job():
-    num_running_jobs = await db.job_count_running()
+    num_running_jobs = await db_jobs.job_count_running()
     if num_running_jobs > 0:
         return {"message": "A job is already running"}
-    nextjob = await db.jobs_get_next_queued_job()
+    nextjob = await db_jobs.jobs_get_next_queued_job()
     if nextjob:
         print(f"Starting Next Job in Queue: {nextjob}")
         print("Starting job: " + str(nextjob["id"]))
         job_config = json.loads(nextjob["job_data"])
         experiment_id = nextjob["experiment_id"]
-        data = await db.experiment_get(experiment_id)
+        data = await db_jobs.experiment_get(experiment_id)
         if data is None:
             # mark the job as failed
-            await db.job_update_status(nextjob["id"], "FAILED")
+            await db_jobs.job_update_status(nextjob["id"], "FAILED")
             return {"message": f"Experiment {id} does not exist"}
         # config = json.loads(data["config"])
 
@@ -79,25 +80,25 @@ async def start_next_job():
 async def stop_job(job_id: str):
     # The way a job is stopped is simply by adding "stop: true" to the job_data
     # This will be checked by the plugin as it runs
-    await db.job_stop(job_id)
+    await db_jobs.job_stop(job_id)
     return {"message": "OK"}
 
 
 @router.get("/delete_all")
 async def job_delete_all():
-    await db.job_delete_all()
+    await db_jobs.job_delete_all()
     return {"message": "OK"}
 
 
 @router.get("/{job_id}")
 async def get_training_job(job_id: str):
-    return await db.job_get(job_id)
+    return await db_jobs.job_get(job_id)
 
 
 @router.get("/{job_id}/output")
 async def get_training_job_output(job_id: str, sweeps: bool = False):
     # First get the template Id from this job:
-    job = await db.job_get(job_id)
+    job = await db_jobs.job_get(job_id)
 
     job_data = json.loads(job["job_data"])
 
@@ -113,7 +114,7 @@ async def get_training_job_output(job_id: str, sweeps: bool = False):
 
     template_id = job_data["template_id"]
     # Then get the template:
-    template = await db.get_training_template(template_id)
+    template = await db_jobs.get_training_template(template_id)
     # Then get the plugin name from the template:
 
     template_config = json.loads(template["config"])
@@ -125,7 +126,7 @@ async def get_training_job_output(job_id: str, sweeps: bool = False):
     # Now we need the current experiment id from the job:
     # experiment_id = job["experiment_id"]
     # Then get the experiment name:
-    # experiment = await db.experiment_get(experiment_id)
+    # experiment = await db_jobs.experiment_get(experiment_id)
     # experiment_name = experiment["name"]
 
     # Now we can get the output.txt from the plugin which is stored in
@@ -141,7 +142,7 @@ async def get_training_job_output(job_id: str, sweeps: bool = False):
 
 @router.get("/template/{template_id}")
 async def get_training_template(template_id: str):
-    return await db.get_training_template(template_id)
+    return await db_jobs.get_training_template(template_id)
 
 
 @router.put("/template/update")
@@ -151,7 +152,7 @@ async def update_training_template(
     try:
         configObject = json.loads(config)
         datasets = configObject["dataset_name"]
-        await db.update_training_template(template_id, name, description, type, datasets, config)
+        await db_jobs.update_training_template(template_id, name, description, type, datasets, config)
     except JSONDecodeError as e:
         logging.error(f"JSON decode error: {e}")
         return {"status": "error", "message": "An error occurred while processing the request."}
@@ -163,7 +164,7 @@ async def update_training_template(
 
 @router.get("/{job_id}/stream_output")
 async def stream_job_output(job_id: str, sweeps: bool = False):
-    job = await db.job_get(job_id)
+    job = await db_jobs.job_get(job_id)
     job_data = job["job_data"]
 
     job_id = secure_filename(job_id)
@@ -214,7 +215,7 @@ async def stream_detailed_json_report(job_id: str, file_name: str):
 
 @router.get("/{job_id}/get_additional_details")
 async def stream_job_additional_details(job_id: str, task: str = "view"):
-    job = await db.job_get(job_id)
+    job = await db_jobs.job_get(job_id)
     job_data = job["job_data"]
     file_path = job_data["additional_output_path"]
     if file_path.endswith(".csv"):
@@ -245,7 +246,7 @@ async def stream_job_additional_details(job_id: str, task: str = "view"):
 
 @router.get("/{job_id}/get_figure_json")
 async def get_figure_path(job_id: str):
-    job = await db.job_get(job_id)
+    job = await db_jobs.job_get(job_id)
     job_data = job["job_data"]
     file_path = job_data.get("plot_data_path", None)
 
@@ -258,7 +259,7 @@ async def get_figure_path(job_id: str):
 
 @router.get("/{job_id}/get_generated_dataset")
 async def get_generated_dataset(job_id: str):
-    job = await db.job_get(job_id)
+    job = await db_jobs.job_get(job_id)
     # Get experiment name
     job_data = job["job_data"]
 
@@ -283,7 +284,7 @@ async def get_generated_dataset(job_id: str):
 @router.get("/{job_id}/get_eval_images")
 async def get_eval_images(job_id: str):
     """Get list of evaluation images for a job"""
-    job = await db.job_get(job_id)
+    job = await db_jobs.job_get(job_id)
     job_data = job["job_data"]
 
     # Check if the job has eval_images_dir
@@ -328,7 +329,7 @@ async def get_eval_images(job_id: str):
 @router.get("/{job_id}/image/{filename}")
 async def get_eval_image(job_id: str, filename: str):
     """Serve individual evaluation image files"""
-    job = await db.job_get(job_id)
+    job = await db_jobs.job_get(job_id)
     job_data = job["job_data"]
 
     # Check if the job has eval_images_dir
